refactor: extract Radiator.Feed.Generator
